UncertainHistogramming

Documentation for UncertainHistogramming.

Base.show

StatsBase.kurtosis

StatsBase.skewness

UncertainHistogramming._online_mean

UncertainHistogramming._update_moment!

UncertainHistogramming._update_moments!

Base.show — Method

show([::IO,] ::GaussianHistogram)
show(::GaussianHistogram)

print the relevant information for a GaussianHistogram.

StatsBase.kurtosis — Function

kurtosis(::GaussianHistogram [, excess = true ])

Pearson (excess) kurtosis of a GaussianHistogram.

Note

For comparison purposes, this kurtosis definition, with excess == true, applied to a Gaussian distribution yields 0. If excess == false, then the kurtosis for a Gaussian is 3.

StatsBase.skewness — Method

skewness(::GaussianHistogram)

Fisher's skewness of a GaussianHistogram.

Note

For comparison purposes, this skewness definition applied to a Gaussian distribution yields identically zero.

UncertainHistogramming._online_mean — Method

_online_mean(xnew, μold, nold)

Update the old mean μold over nold elements with the new data point xnew. This function implements

\[\mu_{n+1} = \mu_n + \frac{x_{n+1} - \mu_{n}}{n+1}.\]

UncertainHistogramming._update_moment! — Method

_update_moment!(::GaussianHistogram, moment_t, μ, σ)

Update the GaussianHistogram's moment_t using an _online_mean with the inclusion of the value-error pair (μ, σ).

UncertainHistogramming._update_moments! — Method

_update_moments!(::GaussianHistogram, μ, σ)

Update the the non-central moments of the GaussianHistogram online with the new value-error pair (μ, σ). This is done because each is non-central moment is the mean non-central moments of each GaussianDistribution.